Integration Workflow

Integrate multiple seurat objects and save to file

integration_workflow(
  batches,
  excluded_cells,
  resolution = seq(0.2, 2, by = 0.2),
  experiment_name = "default_experiment",
  organism = "human",
  ...
)

Arguments

batches
seurat objects for each all batches provided as a list. If named, the resulting integrated object will be identified with corresponding values in 'batch' metadata

excluded_cells
named list of cells to exclude

resolution
value(s) to control the clustering resolution via Seurat::FindMarkers

experiment_name
arbitrary name to identify experiment

organism
either "human" or "mouse"

...

Value

Examples

batches <- seurat_pancreas_reduced %>%
  purrr::map(Seurat::SplitObject, split.by = "dataset") %>%
  purrr::transpose()

inegrated_seu <- integration_workflow(batches)

#> [17:17:29] Initializing seuratTools object...

#> [17:17:29] Initializing seuratTools object...

#> Centering and scaling data matrix

#> Centering and scaling data matrix

#> Computing 2000 integration features

#> Scaling features for provided objects

#> Finding all pairwise anchors

#> Running CCA

#> Merging objects

#> Finding neighborhoods

#> Finding anchors

#> 	Found 2796 anchors

#> Filtering anchors

#> 	Retained 1439 anchors

#> Extracting within-dataset neighbors

#> Merging dataset 2 into 1

#> Extracting anchors for merged samples

#> Finding integration vectors

#> Finding integration vector weights

#> Integrating data

#> Warning: Adding a command log without an assay associated with it

#> PC_ 1 
#> Positive:  CRYBA2, SCGB2A1, G6PC2, VGF, TM4SF4, RGS4, KCNH2, FAM159B, RGS9, ANXA6 
#> 	   KCTD12, PCSK1, RASD1, ADAMTSL2, CD99L2, NR0B1, DHRS2, PEG10, LOXL4, INS 
#> 	   CRH, TUBA4A, PAPPA2, PLCE1, C1orf127, ELMO1, C10orf10, TMEM86B, FAM167A, NPTX2 
#> Negative:  IFITM3, LGALS3, RHOC, SERPING1, S100A11, CDC42EP1, TACSTD2, CTSH, KRT7, TMSB4X 
#> 	   ANXA4, PRSS8, CLDN1, LITAF, LCN2, ZFP36L1, SERPINA3, IFITM2, SDC4, TPM1 
#> 	   SLC44A2, KRT19, ANXA2P2, LAD1, DHRS3, CFB, PPAP2C, TSPAN15, SERPINA5, TM4SF1 
#> PC_ 2 
#> Positive:  KRT8, ELF3, CLDN4, LCN2, KRT18, CFB, PRSS8, SERPINA3, TACSTD2, SDC4 
#> 	   GATM, SLC44A4, KRT7, LAD1, SERPINA5, ABCC3, ANXA4, PIGR, CLDN10, GJB1 
#> 	   PDZK1IP1, SERINC2, KRT19, CLDN1, MUC1, GPX2, AMBP, TSPAN8, SLC4A4, CD74 
#> Negative:  SPARC, COL1A2, BGN, COL3A1, PDGFRB, COL15A1, NID1, COL6A2, COL5A1, COL1A1 
#> 	   MRC2, COL4A1, SFRP2, MMP2, IGFBP4, LRRC32, TIMP3, CYGB, LUM, AEBP1 
#> 	   F2R, COL6A3, DCN, CDH11, MXRA8, PXDN, THBS2, COL5A2, THY1, CTHRC1 
#> PC_ 3 
#> Positive:  IGFBP7, MCAM, PLVAP, TINAGL1, RGCC, TM4SF18, PECAM1, PCAT19, ESAM, CCND1 
#> 	   FLT1, ELTD1, CA2, CFTR, TMEM173, CXCR4, KDR, PMEPA1, VTCN1, CLEC14A 
#> 	   ECSCR, S1PR1, AQP1, MMP7, ACVRL1, SPP1, CDH5, NRP1, ERG, MMRN2 
#> Negative:  CTRB1, CTRB2, CPA2, CTRC, PRSS3P2, REG1B, PNLIP, PNLIPRP2, PRSS1, PLA2G1B 
#> 	   CPA1, PRSS3, CELA2A, RARRES2, PNLIPRP1, BCAT1, CELA3A, KLK1, DPEP1, PDIA2 
#> 	   REG3A, CPB1, CELA3B, ALB, SLC39A5, CEL, GSTA2, FAM129A, REG3G, SPINK1 
#> PC_ 4 
#> Positive:  COL3A1, COL1A2, SFRP2, MXRA8, COL1A1, LUM, PCOLCE, NOTCH3, PDGFRB, COL5A1 
#> 	   COL6A3, DCN, THBS2, VCAN, COL6A1, CYGB, CDH11, COL5A2, SPON2, GGT5 
#> 	   TIMP1, C1S, BGN, ITGA11, MYL9, EDNRA, LAMC3, CTSK, FN1, GPNMB 
#> Negative:  PLVAP, PODXL, PECAM1, RGCC, PCAT19, FLT1, ELTD1, ESAM, KDR, ECSCR 
#> 	   S1PR1, CLEC14A, CDH5, ERG, MMRN2, GPR4, CXCR4, GPR116, TIE1, CD93 
#> 	   ABI3, DYSF, MYCT1, VWF, ESM1, ACVRL1, PASK, RBP5, F2RL3, GMFG 
#> PC_ 5 
#> Positive:  TM4SF4, LOXL4, FAP, CRYBA2, KCTD12, VGF, DPP4, RGS4, PLCE1, PAPPA2 
#> 	   IGFBP2, CRH, NPNT, MUC13, ARRDC4, SERPINA1, PDK4, TMEM45B, SYNDIG1, GPC6 
#> 	   EGFL7, PTP4A3, TMEM86B, CD99L2, SLC40A1, LBH, STC2, IGFBP7, C10orf10, KLHL41 
#> Negative:  HADH, INS, NPTX2, IAPP, C1orf127, PCSK1, WSCD2, CASR, DLK1, ADCYAP1 
#> 	   PFKFB2, RBP4, MEG3, CYYR1, GPM6A, IGSF1, BMP5, TMEM37, FAM159B, ENTPD3 
#> 	   LINC01099, VAT1L, TSPAN1, ASB9, RGS16, FFAR4, SIX3, GCGR, NDRG4, ELMO1 

#> Warning: The default method for RunUMAP has changed from calling Python UMAP via reticulate to the R-native UWOT using the cosine metric
#> To use Python UMAP via reticulate, set umap.method to 'umap-learn' and metric to 'correlation'
#> This message will be shown once per session

#> 17:18:01 UMAP embedding parameters a = 0.9922 b = 1.112

#> 17:18:01 Read 3032 rows and found 30 numeric columns

#> 17:18:01 Using Annoy for neighbor search, n_neighbors = 30

#> 17:18:01 Building Annoy index with metric = cosine, n_trees = 50

#> 0%   10   20   30   40   50   60   70   80   90   100%

#> [----|----|----|----|----|----|----|----|----|----|

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> 

#> |

#> 17:18:02 Writing NN index file to temp file /tmp/RtmpUWXAwF/file3cfb70cc4edb

#> 17:18:02 Searching Annoy index using 1 thread, search_k = 3000

#> 17:18:02 Annoy recall = 100%

#> 17:18:03 Commencing smooth kNN distance calibration using 1 thread

#> 17:18:03 Initializing from normalized Laplacian + noise

#> 17:18:03 Commencing optimization for 500 epochs, with 125324 positive edges

#> 17:18:12 Optimization finished

#> Computing nearest neighbor graph

#> Computing SNN

#> clustering at 0.2 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.9503
#> Number of communities: 8
#> Elapsed time: 0 seconds

#> clustering at 0.4 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.9137
#> Number of communities: 9
#> Elapsed time: 0 seconds

#> clustering at 0.6 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.8847
#> Number of communities: 12
#> Elapsed time: 0 seconds

#> clustering at 0.8 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.8602
#> Number of communities: 14
#> Elapsed time: 0 seconds

#> clustering at 1 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.8389
#> Number of communities: 16
#> Elapsed time: 0 seconds

#> clustering at 1.2 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.8215
#> Number of communities: 19
#> Elapsed time: 0 seconds

#> clustering at 1.4 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.8078
#> Number of communities: 19
#> Elapsed time: 0 seconds

#> clustering at 1.6 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.7938
#> Number of communities: 19
#> Elapsed time: 0 seconds

#> clustering at 1.8 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.7799
#> Number of communities: 19
#> Elapsed time: 0 seconds

#> clustering at 2 resolution

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 3032
#> Number of edges: 100253
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.7665
#> Number of communities: 21
#> Elapsed time: 0 seconds

#> [17:18:33] Get enriched pathway for clusters...

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> Adding missing grouping variables: `group`

#> [17:19:15] 1516 pathways passed the thresholds across all clusters and databases.

#> [17:19:15] Initializing seuratTools object...